README.md
This project was created to be an example for my Intro To Programming Logic and Design students. It is not intended to be a fully functional program, merely a suggestion of what the project should include as a bare minimum. Students are encouraged to elaborate and enhance the project for a better grade.